Move memex.resources to h.resources
This commit moves the AnnotationResource and AnnotationResourceFactory
classes from memex.resources to the existing h.resources module.

This is part of a project to reintegrate the `memex` package into the
main `h` package: https://notes.wtk.io/2017/03/08/reintegrating-memex.
