+++ h/routes.py
@@ -52,7 +52,7 @@ def includeme(config):
     # Annotations & stream
     config.add_route('annotation',
                      '/a/{id}',
-                     factory='memex.resources:AnnotationResourceFactory',
+                     factory='h.resources:AnnotationResourceFactory',
                      traverse='/{id}')
     config.add_route('stream', '/stream')
     config.add_route('stream.user_query', '/u/{user}')
@@ -71,19 +71,19 @@ def includeme(config):
     config.add_route('api.annotations', '/api/annotations')
     config.add_route('api.annotation',
                      '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}',
-                     factory='memex.resources:AnnotationResourceFactory',
+                     factory='h.resources:AnnotationResourceFactory',
                      traverse='/{id}')
     config.add_route('api.annotation_flag',
                      '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}/flag',
-                     factory='memex.resources:AnnotationResourceFactory',
+                     factory='h.resources:AnnotationResourceFactory',
                      traverse='/{id}')
     config.add_route('api.annotation_hide',
                      '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}/hide',
-                     factory='memex.resources:AnnotationResourceFactory',
+                     factory='h.resources:AnnotationResourceFactory',
                      traverse='/{id}')
     config.add_route('api.annotation.jsonld',
                      '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}.jsonld',
-                     factory='memex.resources:AnnotationResourceFactory',
+                     factory='h.resources:AnnotationResourceFactory',
                      traverse='/{id}')
     config.add_route('api.profile', '/api/profile')
     config.add_route('api.debug_token', '/api/debug-token')
