+++ tests/h/routes_test.py
@@ -51,7 +51,7 @@ def test_includeme():
         call('admin_users_activate', '/admin/users/activate'),
         call('admin_users_delete', '/admin/users/delete'),
         call('admin_users_rename', '/admin/users/rename'),
-        call('annotation', '/a/{id}', factory='memex.resources:AnnotationResourceFactory', traverse='/{id}'),
+        call('annotation', '/a/{id}', factory='h.resources:AnnotationResourceFactory', traverse='/{id}'),
         call('stream', '/stream'),
         call('stream.user_query', '/u/{user}'),
         call('stream.tag_query', '/t/{tag}'),
@@ -60,19 +60,19 @@ def test_includeme():
         call('api.annotations', '/api/annotations'),
         call('api.annotation',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}',
-             factory='memex.resources:AnnotationResourceFactory',
+             factory='h.resources:AnnotationResourceFactory',
              traverse='/{id}'),
         call('api.annotation_flag',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}/flag',
-             factory='memex.resources:AnnotationResourceFactory',
+             factory='h.resources:AnnotationResourceFactory',
              traverse='/{id}'),
         call('api.annotation_hide',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}/hide',
-             factory='memex.resources:AnnotationResourceFactory',
+             factory='h.resources:AnnotationResourceFactory',
              traverse='/{id}'),
         call('api.annotation.jsonld',
              '/api/annotations/{id:[A-Za-z0-9_-]{20,22}}.jsonld',
-             factory='memex.resources:AnnotationResourceFactory',
+             factory='h.resources:AnnotationResourceFactory',
              traverse='/{id}'),
         call('api.profile', '/api/profile'),
         call('api.debug_token', '/api/debug-token'),
